backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
attrs==25.3.0
bcrypt==4.3.0
Brotli==1.1.0
certifi==2025.7.9
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cssselect2==0.8.0
cycler==0.12.1
debugpy==1.8.12
decorator==5.1.1
distro==1.9.0
dnspython==2.7.0
docopt==0.6.2
ecdsa==0.19.1
email_validator==2.2.0
executing==2.2.0
fastapi==0.116.1
fonttools==4.56.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
ipykernel==6.29.5
ipython==8.31.0
jedi==0.19.2
jiter==0.10.0
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
markdown2==2.5.3
matplotlib==3.10.0
matplotlib-inline==0.1.7
md2pdf==1.0.1
mido==1.3.3
nest-asyncio==1.6.0
numpy==2.2.3
openai==1.95.0
packaging==24.2
parso==0.8.4
pdfkit==1.0.0
pillow==10.4.0
platformdirs==4.3.6
pluggy==1.6.0
pretty_midi==0.2.10
prompt_toolkit==3.0.50
psutil==6.1.1
pure_eval==0.2.3
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydyf==0.11.0
Pygments==2.19.1
PyGuitarPro==0.9.3
PyMuPDF==1.26.3
pyparsing==3.2.1
pyphen==0.17.2
pytest==8.4.1
pytest-asyncio==1.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
python-rtmidi==1.5.8
pywin32==308
pyzmq==26.2.0
requests==2.32.4
rsa==4.9.1
setuptools==78.1.0
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
starlette==0.47.2
tinycss2==1.4.0
tinyhtml5==2.0.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
wcwidth==0.2.13
weasyprint==65.1
webencodings==0.5.1
zopfli==0.2.3.post1
